requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
